+++ core/src/main/java/feign/Contract.java
@@ -26,7 +26,6 @@ import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import javax.inject.Named;
 
 /** Defines what annotations and values are valid on interfaces. */
 public interface Contract {
@@ -177,16 +176,10 @@ public interface Contract {
       boolean isHttpAnnotation = false;
       for (Annotation annotation : annotations) {
         Class<? extends Annotation> annotationType = annotation.annotationType();
-        if (annotationType == Param.class || annotationType == Named.class) {
-          String name =
-              annotationType == Param.class
-                  ? ((Param) annotation).value()
-                  : ((Named) annotation).value();
+        if (annotationType == Param.class) {
+          String name = ((Param) annotation).value();
           checkState(
-              emptyToNull(name) != null,
-              "%s annotation was empty on param %s.",
-              annotationType.getSimpleName(),
-              paramIndex);
+              emptyToNull(name) != null, "Param annotation was empty on param %s.", paramIndex);
           nameParam(data, name, paramIndex);
           if (annotationType == Param.class) {
             Class<? extends Param.Expander> expander = ((Param) annotation).expander();