+++ core/src/test/java/feign/LoggerTest.java
@@ -239,7 +239,6 @@ public class LoggerTest {
 
     @Test
     public void unknownHostEmits() throws IOException, InterruptedException {
-
       SendsStuff api =
           Feign.builder()
               .logger(logger)
@@ -289,7 +288,6 @@ public class LoggerTest {
 
     @Test
     public void retryEmits() throws IOException, InterruptedException {
-
       thrown.expect(FeignException.class);
 
       SendsStuff api =