+++ core/src/test/java/feign/client/DefaultClientTest.java
@@ -24,7 +24,6 @@ import static org.junit.Assert.assertEquals;
 import com.squareup.okhttp.mockwebserver.MockResponse;
 import com.squareup.okhttp.mockwebserver.SocketPolicy;
 import com.squareup.okhttp.mockwebserver.rule.MockWebServerRule;
-import dagger.Lazy;
 import feign.Client;
 import feign.Feign;
 import feign.FeignException;
@@ -35,9 +34,7 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.net.ProtocolException;
 import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLSession;
-import javax.net.ssl.SSLSocketFactory;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -109,20 +106,7 @@ public class DefaultClientTest {
     api.patch();
   }
 
-  Client trustSSLSockets =
-      new Client.Default(
-          new Lazy<SSLSocketFactory>() {
-            @Override
-            public SSLSocketFactory get() {
-              return TrustingSSLSocketFactory.get();
-            }
-          },
-          new Lazy<HostnameVerifier>() {
-            @Override
-            public HostnameVerifier get() {
-              return HttpsURLConnection.getDefaultHostnameVerifier();
-            }
-          });
+  Client trustSSLSockets = new Client.Default(TrustingSSLSocketFactory.get(), null);
 
   @Test
   public void canOverrideSSLSocketFactory() throws IOException, InterruptedException {
@@ -139,21 +123,11 @@ public class DefaultClientTest {
 
   Client disableHostnameVerification =
       new Client.Default(
-          new Lazy<SSLSocketFactory>() {
+          TrustingSSLSocketFactory.get(),
+          new HostnameVerifier() {
             @Override
-            public SSLSocketFactory get() {
-              return TrustingSSLSocketFactory.get();
-            }
-          },
-          new Lazy<HostnameVerifier>() {
-            @Override
-            public HostnameVerifier get() {
-              return new HostnameVerifier() {
-                @Override
-                public boolean verify(String s, SSLSession sslSession) {
-                  return true;
-                }
-              };
+            public boolean verify(String s, SSLSession sslSession) {
+              return true;
             }
           });