+++ jackson/src/main/java/feign/jackson/JacksonEncoder.java
@@ -17,20 +17,27 @@ package feign.jackson;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import feign.RequestTemplate;
 import feign.codec.EncodeException;
 import feign.codec.Encoder;
+import java.util.Collections;
 
 public class JacksonEncoder implements Encoder {
   private final ObjectMapper mapper;
 
   public JacksonEncoder() {
+    this(Collections.<Module>emptyList());
+  }
+
+  public JacksonEncoder(Iterable<Module> modules) {
     this(
         new ObjectMapper()
             .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-            .configure(SerializationFeature.INDENT_OUTPUT, true));
+            .configure(SerializationFeature.INDENT_OUTPUT, true)
+            .registerModules(modules));
   }
 
   public JacksonEncoder(ObjectMapper mapper) {