+++ jaxb/src/main/java/feign/jaxb/JAXBDecoder.java
@@ -21,7 +21,6 @@ import feign.codec.DecodeException;
 import feign.codec.Decoder;
 import java.io.IOException;
 import java.lang.reflect.Type;
-import javax.inject.Inject;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Unmarshaller;
 
@@ -47,7 +46,6 @@ import javax.xml.bind.Unmarshaller;
 public class JAXBDecoder implements Decoder {
   private final JAXBContextFactory jaxbContextFactory;
 
-  @Inject
   public JAXBDecoder(JAXBContextFactory jaxbContextFactory) {
     this.jaxbContextFactory = jaxbContextFactory;
   }