+++ jaxb/src/test/java/feign/jaxb/JAXBCodecTest.java
@@ -19,47 +19,18 @@ import static feign.Util.UTF_8;
 import static feign.assertj.FeignAssertions.assertThat;
 import static org.junit.Assert.assertEquals;
 
-import dagger.Module;
-import dagger.ObjectGraph;
 import feign.RequestTemplate;
 import feign.Response;
-import feign.codec.Decoder;
 import feign.codec.Encoder;
 import java.util.Collection;
 import java.util.Collections;
-import javax.inject.Inject;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import org.junit.Test;
 
-public class JAXBModuleTest {
-  @Module(includes = JAXBModule.class, injects = EncoderAndDecoderBindings.class)
-  static class EncoderAndDecoderBindings {
-    @Inject Encoder encoder;
-
-    @Inject Decoder decoder;
-  }
-
-  @Module(includes = JAXBModule.class, injects = EncoderBindings.class)
-  static class EncoderBindings {
-    @Inject Encoder encoder;
-  }
-
-  @Module(includes = JAXBModule.class, injects = DecoderBindings.class)
-  static class DecoderBindings {
-    @Inject Decoder decoder;
-  }
-
-  @Test
-  public void providesEncoderDecoder() throws Exception {
-    EncoderAndDecoderBindings bindings = new EncoderAndDecoderBindings();
-    ObjectGraph.create(bindings).inject(bindings);
-
-    assertEquals(JAXBEncoder.class, bindings.encoder.getClass());
-    assertEquals(JAXBDecoder.class, bindings.decoder.getClass());
-  }
+public class JAXBCodecTest {
 
   @XmlRootElement
   @XmlAccessorType(XmlAccessType.FIELD)
@@ -84,14 +55,11 @@ public class JAXBModuleTest {
 
   @Test
   public void encodesXml() throws Exception {
-    EncoderBindings bindings = new EncoderBindings();
-    ObjectGraph.create(bindings).inject(bindings);
-
     MockObject mock = new MockObject();
     mock.value = "Test";
 
     RequestTemplate template = new RequestTemplate();
-    bindings.encoder.encode(mock, template);
+    new JAXBEncoder(new JAXBContextFactory.Builder().build()).encode(mock, template);
 
     assertThat(template)
         .hasBody(
@@ -104,8 +72,7 @@ public class JAXBModuleTest {
     JAXBContextFactory jaxbContextFactory =
         new JAXBContextFactory.Builder().withMarshallerJAXBEncoding("UTF-16").build();
 
-    JAXBModule jaxbModule = new JAXBModule(jaxbContextFactory);
-    Encoder encoder = jaxbModule.encoder(new JAXBEncoder(jaxbContextFactory));
+    Encoder encoder = new JAXBEncoder(jaxbContextFactory);
 
     MockObject mock = new MockObject();
     mock.value = "Test";
@@ -126,8 +93,7 @@ public class JAXBModuleTest {
             .withMarshallerSchemaLocation("http://apihost http://apihost/schema.xsd")
             .build();
 
-    JAXBModule jaxbModule = new JAXBModule(jaxbContextFactory);
-    Encoder encoder = jaxbModule.encoder(new JAXBEncoder(jaxbContextFactory));
+    Encoder encoder = new JAXBEncoder(jaxbContextFactory);
 
     MockObject mock = new MockObject();
     mock.value = "Test";
@@ -150,8 +116,7 @@ public class JAXBModuleTest {
             .withMarshallerNoNamespaceSchemaLocation("http://apihost/schema.xsd")
             .build();
 
-    JAXBModule jaxbModule = new JAXBModule(jaxbContextFactory);
-    Encoder encoder = jaxbModule.encoder(new JAXBEncoder(jaxbContextFactory));
+    Encoder encoder = new JAXBEncoder(jaxbContextFactory);
 
     MockObject mock = new MockObject();
     mock.value = "Test";
@@ -172,8 +137,7 @@ public class JAXBModuleTest {
     JAXBContextFactory jaxbContextFactory =
         new JAXBContextFactory.Builder().withMarshallerFormattedOutput(true).build();
 
-    JAXBModule jaxbModule = new JAXBModule(jaxbContextFactory);
-    Encoder encoder = jaxbModule.encoder(new JAXBEncoder(jaxbContextFactory));
+    Encoder encoder = new JAXBEncoder(jaxbContextFactory);
 
     MockObject mock = new MockObject();
     mock.value = "Test";
@@ -199,9 +163,6 @@ public class JAXBModuleTest {
 
   @Test
   public void decodesXml() throws Exception {
-    DecoderBindings bindings = new DecoderBindings();
-    ObjectGraph.create(bindings).inject(bindings);
-
     MockObject mock = new MockObject();
     mock.value = "Test";
 
@@ -213,6 +174,8 @@ public class JAXBModuleTest {
         Response.create(
             200, "OK", Collections.<String, Collection<String>>emptyMap(), mockXml, UTF_8);
 
-    assertEquals(mock, bindings.decoder.decode(response, MockObject.class));
+    JAXBDecoder decoder = new JAXBDecoder(new JAXBContextFactory.Builder().build());
+
+    assertEquals(mock, decoder.decode(response, MockObject.class));
   }
 }